Updated the `articlesScreen` function in `ArticlesNavigation` to accept a lambda with no parameters for navigating to the details screen . This change simplifies the navigation logic and removes the need to pass an article ID.

+++ feature/home/src/main/java/com/droidos/home/presentation/navigation/ArticlesNavigation.kt
@@ -8,7 +8,7 @@ import com.droidos.home.presentation.ArticlesRoute
 const val ArticleRoute = "articlesScreen"
 
 fun NavGraphBuilder.articlesScreen(
-    onNavToDetails: (Int) -> Unit,
+    onNavToDetails: () -> Unit,
 ) {
     composable(ArticleRoute) {
         ArticlesRoute(onNavToDetails)
